Moved report configuration to the story
